Move consumer pact test and related docker files to pact package
